adição código edifício

+++ QAI.py
@@ -4,6 +4,9 @@ create_top(big_text_title='QAI em Escritórios', subtitle='Design e organizaçã
 
 st.info('Para garantir a autenticidade da pesquisa, será enviado um código de verificação para seu email. **Suas respostas são anônimas e confidenciais.**', icon='ℹ️')
 st.title('')
+
+col1, col2, col3 = st.columns(3)
+edificio = col2.text_input(label='None', label_visibility="collapsed", max_chars=8, placeholder="Código do edifício")
 caixa_email, botao_enviar = st.columns([1,0.1])
 mail = caixa_email.text_input('no label', label_visibility='hidden', max_chars=100, key='emaildoparticipante', placeholder='Insira seu email')
 botao_enviar.write('')
@@ -26,12 +29,13 @@ confirmacao_texto = meio.text_input('nolabel', label_visibility='hidden', max_ch
 
 st.title('')
 if centered_button('Iniciar questionário'):
-    if confirmacao_texto:
+    if confirmacao_texto and edificio:
         if st.session_state['auth']:
             st.write(st.session_state['auth']) 
             if confirmacao_texto != st.session_state['auth']:
                 st.error('Usuário não autorizado / Código incorreto')
             else:
+                PeR['edificio'] = edificio
                 open(r'base/hierarquia.txt', 'w').write('')
                 open(r'base/person_mail.txt', 'w').write('')
                 open(r'base/person_mail.txt', 'w').write(mail)
@@ -39,4 +43,4 @@ if centered_button('Iniciar questionário'):
         else:
             st.error('Código de verificação não gerado, insira seu email, clique no botão de envio e copie o código que receber por email')
     else:
-        st.error('Código de verificação não inserido')
+        st.error('Insira todas as informações para prosseguir')
